Refactor the rest of containers into components
- Delete OpenIssuesContainer bc it was already in the state for the corresponding component
- Move SignupContainer but kept it as the container bc Signup is already super long
- Refactor VolunteersContainer and VolunteerSignupContainer
