--- app/containers/VolunteersContainer.jsx
@@ -1,10 +1,0 @@
-import { connect } from 'react-redux'
-import Volunteers from '../components/Volunteers'
-
-function mapStateToProps (state) {
-  return {
-    volunteers: state.volunteers
-  }
-}
-
-export default connect(mapStateToProps)(Volunteers)
